Windows support for salmonella (popen)
On unix, apply `qs' to the command.  Windows doesn't seem to play well
with the result of `qs' on commands.

Explicitly quote Scheme expressions given as arguments to csi.

Handle the output of close-input-pipe on Windows.

salmonella-epidemy is still not supported, since it still uses
process-fork (not available on Windows).