Set CHICKEN_PREFIX
This fixes a bug related to the fact that (chicken-home) is in the
include path.  (chicken-home) would pick C_INSTALL_SHARE_HOME if
CHICKEN_PREFIX is not set.

Some eggs like setup-helper and check-errors install files under
(chicken-home), and that causes problems because those files are not
removed by clear-repo!. So, eggs that depend on setup-helper, for
example, but not specify it on their dependencies list, may be
successfully installed if setup-helper had been previously installed.

We fix that by copying setup.defaults to salmonella's private
repository and setting CHICKEN_PREFIX.  By setting CHICKEN_PREFIX,
(chicken-home) will now point to salmonella's private repository, and
will be removed when clear-repo! is called.