Better handling of test dependencies
`test-egg' was ignoring fetch and installation of test dependencies
and was calling the test script even in case of a fetch/installation
failure of test dependencies.  It also would only return a single
report object representing the result of tests.

Now `test-dep' generates a report object for each test dependency, for
both actions: test and install.  It now returns all the reports as a
list.  In case of any fetch or installation failure of test
dependencies, the test script is not executed and a list of report
objects is returned.  The report object for the test action of the egg
contains information on the test dependency whose fetch or
installation failed.

The log format version has been bumped to 3, as the format changed.
Now we have log data for the test dependencies.  The action field for
test dependencies is a list in the format (<action> <egg>), where
<action> can be fetch-test-dep or install-test-dep, and <egg> is the
egg which depends on the test dependency.