salmonella-log-parser: log-eggs: only consider eggs given as inputs to salmonella
5fda3fafbeb0300 adds new metadata to make it possible to identify
errors in the installation of test dependencies -- new log entries are
added to the log file, and those would be considered as "eggs" by
`log-eggs', as it would list all eggs in the log, except skipped ones.
That's a problem.

This change makes log-eggs only list eggs that were explicitly given
as inputs to salmonella.  Intermediate eggs (e.g., test dependencies)
that were not given as inputs to salmonella are not listed by
`log-eggs', as their metadata are not in the log -- that would break
the other salmonella tools (e.g., salmonella-html-report) that rely on
the existence of egg metadata for eggs returned by `log-eggs'.