Remove --eggs-source-dir command line option
--eggs-source-dir requires `gather-egg-information`, which does not
exist in CHICKEN 5.  It's still possible, however, to test eggs
without contacting a web server by configuring chicken-install to use
a local repository.

In CHICKEN 4, use something like:

  $ salmonella --chicken-install-args="-t local -l /home/mario/src/chicken-4-eggs -prefix <repo> -test" egg1 ...

In CHICKEN 5, set up the `location' form in
PREFIX/share/chicken/setup.defaults to point to a directory containing
the source of eggs.

Note that the directory layouts expected by CHICKEN 4 and 5 are
different.

chicken-install for CHICKEN 4 expects:
    <eggs-source-dir>/<egg>/tags/<version>

chicken-install for CHICKEN 5 expects:
    <eggs-source-dir>/<egg>/<version>