Handle "non-standard" CHICKEN tool names
The name of CHICKEN tools can be tweaked at build time via
CSI_PROGRAM, CSC_PROGRAM, CHICKEN_INSTALL_PROGRAM, PROGRAM_PREFIX and
PROGRAM_SUFFIX as configuration parameters for make.

Arch Linux, for example, tweaks CSI_PROGRAM and CSC_PROGRAM:
https://github.com/archlinux/svntogit-community/blob/8032b4cc90ed4d70732ca0d8e25f6a892d4ad0ec/trunk/PKGBUILD

This change makes salmonella handle cases in which the name of CHICKEN
tools were tweaked at build time.

In case --chicken-installation-prefix is used, it can be hard to infer
which names the CHICKEN tools have, so we resort to the "standard"
csi, csc and chicken-install names.  For the cases where tools in the
prefix pointed by --chicken-installation-prefix were tweaked via the
aforementioned options, new parameters for salmonella have been
introduced:

* --csi
* --csc
* --chicken-install

Those parameters are expected to point to the the path to the
respective CHICKEN tools.  They clobber the default names/paths.

Thanks to nikola2 on IRC for pointing out this issue.