end-time: Make sure the record actually represents the end of execution
When salmonella gets killed before it is finished and leaves an
incomplete log behind, the last log record might not represent the end
of the execution.

This change makes `end-time` check whether the action of the last
record is really `end`, and report an error in case it isn't
(corrupted log).

Thanks to Vasilij Schneidermann for reporting the issue.