salmonella.scm: Use the repository path set at installation time
The determination of the repository path can be affected by the value
of the `CHICKEN_INSTALL_REPOSITORY` environment variable, which is
normally set b tools that create "virtual environments" for CHICKEN
(e.g., cenv, basket).

Avoid issues related to that by unsetting `CHICKEN_INSTALL_REPOSITORY`
when determining the repository path.  By doing that we get the
repository path set at CHICKEN installation time, which is the one we
want to copy the basic modules from.

Thanks to Jeremy Steward for reporting the issue and testing this
patch.